Fix division by zero in Item

+++ src/com/poestats/Config.java
@@ -97,7 +97,7 @@ public class Config {
     // Other?
     //------------------------------------------------------------------------------------------------------------
 
-    public static final int item_pricePrecision = 1000;
+    public static final double item_pricePrecision = 1000.0;
     public static final int misc_defaultLeagueLength = 90;
     public static final int monitorTimeoutMS = 500;
     public static final long startTime = System.currentTimeMillis();
